Implemented design for Login and Random Screens.
In Authorizer implemented "reauthorize" variable which defines if this method is called from post card rater or not. When we rate card for the first time of application running - somehow it fails so we need to authorize again. This authorizer saves cookies. But when we authorize from login view - we need to clear all cookies because they are used and result is always success (does not depend on new credentials you entered).